% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskmeans-package.R
\docType{package}
\name{coseq-package}
\alias{coseq-package}
\alias{maskmeans-package}
\title{Multi-view agglomeration/splitting K-means clustering algorithm}
\description{
Brief description of package ...
}
\details{
\tabular{ll}{ Package: \tab maskmeans\cr Type: \tab Package\cr Version:
\tab 0.0.1\cr Date: \tab 2018-03-13\cr License: \tab GPL (>=3)\cr LazyLoad:
\tab yes\cr }
}
\references{
To be completed ...
}
\author{
Andrea Rau, Cathy Maugis-Rabusseau, Antoine Godichon-Baggioni

Maintainer: Andrea Rau <\url{andrea.rau@inra.fr}>
}
\keyword{cluster}
\keyword{models}
