% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregHierarchique.R
\name{Crit2}
\alias{Crit2}
\title{Calculate the value of the criterion}
\usage{
Crit2(X, mv, gamma, weights, cluster)
}
\arguments{
\item{X}{Matrix of multi-view data, where the first view corresponds to the 
principal data used to obtain the partition in \code{cluster.init}}

\item{mv}{Vector corresponding to the size of each data view. The sum of \code{mv} should 
correspond to the number of columns in \code{X}.}

\item{gamma}{??}

\item{weights}{Blah blah}

\item{cluster}{Blah blah}
}
\description{
Calculate the value of the criterion
}
