% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregHierarchiqueProbapost.R
\name{Crit3}
\alias{Crit3}
\title{Title}
\usage{
Crit3(X, mv, gamma, weights, probapost, cluster)
}
\arguments{
\item{X}{}

\item{mv}{}

\item{gamma}{}

\item{weights}{}

\item{probapost}{}

\item{cluster}{}
}
\description{
Title
}
