% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregHierarchiqueProbapost.R
\name{cutreeNewProbapost}
\alias{cutreeNewProbapost}
\title{Title}
\usage{
cutreeNewProbapost(R, K, probapost.init)
}
\arguments{
\item{R}{}

\item{K}{}

\item{probapost.init}{}
}
\description{
Title
}
