% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregHierarchique.R
\name{hmv1}
\alias{hmv1}
\title{Aggregation of hard clusters based on multi-view input}
\usage{
hmv1(X, mv, gamma, cluster.init, weightsopt = TRUE)
}
\arguments{
\item{X}{Matrix of multi-view data, where the first view corresponds to the 
principal data used to obtain the partition in \code{cluster.init}}

\item{mv}{Vector corresponding to the size of each data view. The sum of \code{mv} should 
correspond to the number of columns in \code{X}.}

\item{gamma}{??}

\item{cluster.init}{Vector of available cluster labels}

\item{weightsopt}{If \code{TRUE}, run algorithm in weighted multi-view mode; if FALSE, weights are}
}
\description{
Aggregation of hard clusters based on multi-view input
}
