% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregHierarchiqueProbapost.R
\name{hmvprobapost}
\alias{hmvprobapost}
\title{Aggregation of fuzzy clusters based on multi-view inputs}
\usage{
hmvprobapost(X, mv, gamma, probapost.init)
}
\arguments{
\item{X}{Matrix of multi-view data, where the first view corresponds to the 
principal data used to obtain the partition in \code{cluster.init}}

\item{mv}{Vector corresponding to the size of each data view. The sum of \code{mv} should 
correspond to the number of columns in \code{X}.}

\item{gamma}{??}

\item{probapost.init}{Matrix of available conditional probabilities of cluster membership
for \code{X}}
}
\description{
Aggregation of fuzzy clusters based on multi-view inputs
}
\examples{

}
