% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splitting.R
\name{MAP_smooth}
\alias{MAP_smooth}
\title{Extract the classification by MAP estimator (complete or "smooth")}
\usage{
MAP_smooth(probapost, seuil = 0.8)
}
\arguments{
\item{probapost}{Matrix of posterior probabilities}

\item{seuil}{Threshold, by default 0.80}
}
\value{
Vector of cluster labels. Observations with maximum conditional probabilithy less than \code{threshold} are
assigned a cluster label of 0.
}
\description{
Extract the classification by MAP estimator (complete or "smooth")
}
