% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskmeans-package.R
\docType{package}
\name{maskmeans-package}
\alias{maskmeans-package}
\title{Multi-view agglomeration/splitting K-means clustering algorithm}
\description{
This package incorporates algorithms for aggregating or splitting an 
existing hard or soft classification using multi-view data.
}
\details{
\tabular{ll}{ Package: \tab maskmeans\cr Type: \tab Package\cr Version:
\tab 0.0.15\cr Date: \tab 2021-11-08\cr License: \tab GPL (>=3.3.1)\cr LazyLoad:
\tab yes\cr }
}
\references{
Godichon-Baggioni, A., Maugis-Rabusseau, C. and Rau, A. (2020) Multi-view 
cluster aggregation and splitting, with an application to multi-omic breast 
cancer data. Annals of Applied Statistics, 14:2, 752-767.
}
\author{
Andrea Rau, Cathy Maugis-Rabusseau, Antoine Godichon-Baggioni

Maintainer: Andrea Rau <\url{andrea.rau@inra.fr}>
}
\keyword{cluster}
\keyword{models}
