% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregation.R
\name{maskmeans_cutree}
\alias{maskmeans_cutree}
\title{Cut an aggregation tree from maskmeans output}
\usage{
maskmeans_cutree(obj, K, clustering_init)
}
\arguments{
\item{obj}{An object of class \code{"maskmeans"}}

\item{K}{An integer scalar with the desired number of clusters}

\item{clustering_init}{Initial partition or posterior probabilities used to cluster the multi-view data}
}
\value{
\item{classif }{Vector of group memberships after cutting tree}
\item{probapost }{Matrix of group posterior probabilities after cutting tree}
}
\description{
Cut an aggregation tree into several groups after running \code{maskmeans_aggregation} or
\code{maskmeans(..., type = "aggregation")}.
}
