% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPlot.R
\name{maskmeans_plot}
\alias{maskmeans_plot}
\title{Plot results of the multi-view aggregation/splitting K-means algorithm}
\usage{
maskmeans_plot(obj, type = c("dendrogram", "heights", "weights_line",
  "weights", "criterion", "tree"), ...)
}
\arguments{
\item{obj}{Object of class \code{"maskmeans"} resulting from a call to the \code{maskemans}
function}

\item{type}{Graphic to be produced: one or more of \code{c("dendrogram", "heights", 
"weights_line", "weights", "criterion", "tree", "probapost_boxplot")}.}

\item{...}{Additional optional parameters. In particular, for cluster tree plots, the original
multi-view data should be provided in \code{mv_data}.}
}
\value{
List of one or more ggplot2 objects.
}
\description{
Produce a variety of plots after running the multi-view aggregation or splitting
version of the K-means algorithm.
}
