% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregation.R
\name{mv_aggregation}
\alias{mv_aggregation}
\title{Aggregation of hard clusters based on multi-view input}
\usage{
mv_aggregation(X, mv, cluster_init, gamma = 2, use_mv_weights = TRUE)
}
\arguments{
\item{X}{Matrix of multi-view data, where the first view corresponds to the 
principal data used to obtain the partition in \code{cluster.init}}

\item{mv}{Vector corresponding to the size of each data view. The sum of \code{mv} should 
correspond to the number of columns in \code{X}.}

\item{cluster_init}{Vector of available cluster labels}

\item{gamma}{Parameter that controls the distribution of view weights. Default value is 2.}

\item{use_mv_weights}{If \code{TRUE}, run algorithm in weighted multi-view mode; if FALSE, the
weight for each view is set to be equal.}
}
\value{
\item{merged_clusters }{Matrix providing each pair of merged clusters at each iteration of the algorithm}
\item{hclust }{Object of class "hclust" to be used for plotting cluster agglomerations}
\item{weights }{Matrix of dimension \code{v} x \code{niterations}, where \code{v} is the number of
views and \code{niterations} is the number of successive agglomerations, providing the multi-view weights}
\item{criterion }{Value taken on by the agglomerative criterion at each iteration}
}
\description{
Aggregation of hard clusters based on multi-view input
}
