% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPlot.R
\name{mv_plot}
\alias{mv_plot}
\title{Overview plot of multi-view data}
\usage{
mv_plot(mv_data, scale = TRUE, ...)
}
\arguments{
\item{mv_data}{Multi-view data, either in the form of a concatenated matrix (where the columns of the views have
been combined) or in the form of a list, where each element contains one of the views. In the former case, the argument
\code{mv} should also be provided to indicate the dimension of each of the views (\code{sum(mv)} should be equal to
the number of columns in the matrix). In the latter case, each matrix in the list should have observations (rows) sorted
in the same order, as matrices will be combined by columns within the function; note that \code{mv} will be detected 
automtically based on the number of columns in each matrix.}

\item{scale}{If \code{TRUE}, data will be scaled so that all variables are normalized and views are standardized according 
to their size.}

\item{...}{Additional optional parameters. In particular, if points should be colored by cluster membership, a vector
\code{labels} providing cluster assigments should be provided.}
}
\value{
A multi-facetted plot, with one facet per view. Univariate views are represented as densities, bivariate views
are represented as scatterplots, and multivariate views are represented by plotting scatterplots of the first two 
principal components.
}
\description{
Global graphical representation of multi-view data.
}
