% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatedData.R
\name{mv_simulate}
\alias{mv_simulate}
\title{Simulate multi-view data}
\usage{
mv_simulate(type = "D6", delta, n, K, sigma, sigma_V6 = NULL,
  delta_V6 = NULL)
}
\arguments{
\item{type}{One of \code{D1}, \code{D2}, \code{D3}, \code{D4}, \code{D5}, \code{D6} 
specifying the type of simulated data to generate.}

\item{delta}{(Optional) Multiplicative factor controlling the spread of values around the origin in the first 
view for simulations of type \code{D6}. Defaults to 7.}

\item{n}{(Optional) Number of observations per cluster for simulations of type \code{D6}. Defaults to 250.}

\item{K}{(Optional) Number of clusters in the first view for simulations of type \code{D6}. As this view is made up
of a single cluster at the origin surrounded by evenly spaced clusters in a circular pattern around
it, this number should be odd. Defaults to 9.}

\item{sigma}{(Optional) Variance of noise to be added to views 2 and 4 for simulations of type \code{D6}.
Defaults to 1.}

\item{sigma_V6}{(Optional) If desired, a different sigma parameter for view 6. Defaults to 2*sigma.}

\item{delta_V6}{(Optional) If desired, a different delta parameter for view 6. Defaults to 2*delta.}
}
\value{
\item{data }{Multi-view simulated data}
\item{labels }{Matrix of dimension \code{n} x \code{v}, where \code{n} is the number
of observations and \code{v} the number of views, representing the true labels used
to generate the data}
}
\description{
Generate simulated multi-view data to illustrate the agglomerative and splitting versions of the
multi-view K-means algorithm.
}
\examples{
set.seed(12345)
sim_1 <- mv_simulate(type = "D1")
sim_2 <- mv_simulate(type = "D2")
sim_3 <- mv_simulate(type = "D3")
sim_4 <- mv_simulate(type = "D4")
sim_5 <- mv_simulate(type = "D5")
sim_6a <- mv_simulate(type = "D6")
sim_6b <- mv_simulate(type = "D6", delta=7, n=200, K=5, sigma=0.5)
}
