% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPlot.R
\name{plotprobaseuil}
\alias{plotprobaseuil}
\title{Title}
\usage{
plotprobaseuil(Res, probapost.init, seuil = 0.8)
}
\arguments{
\item{Res}{}

\item{probapost.init}{}

\item{seuil}{}
}
\description{
Title
}
