% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPlot.R
\name{PlotResSplittingbis}
\alias{PlotResSplittingbis}
\title{Title}
\usage{
PlotResSplittingbis(Res, plotCRIT = T, weights_step = F,
  weights_clust = F)
}
\arguments{
\item{Res}{}

\item{plotCRIT}{}

\item{weights_step}{}

\item{weights_clust}{}
}
\description{
Title
}
