% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splitting.R
\name{probapost_K}
\alias{probapost_K}
\title{Extract the posterior probabilities from a specified step in the splitting tree}
\usage{
probapost_K(probapost, ksplit, K)
}
\arguments{
\item{probapost}{Final matrix of posterior probabilities}

\item{ksplit}{Desired level of the splitting tree}

\item{K}{Desired number of clusters}
}
\value{
Matrix of posterior probabilities.
}
\description{
Extract the posterior probabilities from a specified step in the splitting tree
}
