% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPlot.R
\name{split_zoom}
\alias{split_zoom}
\title{Plot zoom results of the final weights from the multi-view splitting K-means algorithm}
\usage{
split_zoom(obj, initial_cluster = 1, final_K = NULL, mv_names = NULL, ...)
}
\arguments{
\item{obj}{Object of class \code{"maskmeans"} resulting from a call to the \code{maskmeans}
function with \code{type = "splitting"} and \code{perCluster_mv_weights = TRUE}.}

\item{initial_cluster}{A numerical value indicating the initial cluster for which splits should be visualized.}

\item{final_K}{Total number of clusters in selected model. By default, the value in \code{obj$final_K} will be used if it exists, and otherwise the 
maximum number of split clusters will be used.}

\item{mv_names}{If desired, a vector of multiview names to be used}

\item{...}{Additional optional parameters.}
}
\value{
A ggplot2 object
}
\description{
Produce a plot of the per-cluster weights from the multi-view splitting K-means algorithm for a specified initial cluster value.
}
