% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplittingClusters.R
\name{splittingClusters}
\alias{splittingClusters}
\title{Title}
\usage{
splittingClusters(X, mv, gamma, Kmax, cluster.init, weightsopt = T,
  testkmeans = T)
}
\arguments{
\item{X}{}

\item{mv}{}

\item{gamma}{}

\item{Kmax}{}

\item{cluster.init}{}

\item{weightsopt}{}

\item{testkmeans}{}
}
\description{
Title
}
