% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplittingClusters.R
\name{splittingClustersbis}
\alias{splittingClustersbis}
\title{Title}
\usage{
splittingClustersbis(X, mv, gamma = 1, Kmax, cluster.init)
}
\arguments{
\item{X}{}

\item{mv}{}

\item{gamma}{}

\item{Kmax}{}

\item{cluster.init}{}
}
\description{
Title
}
