% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregHierarchique.R
\name{weightcalculate}
\alias{weightcalculate}
\title{Calculation of multi-view weights for hard clusters}
\usage{
weightcalculate(X, mv, centers, cluster, gamma)
}
\description{
Calculation of multi-view weights for hard clusters
}
