% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplittingClusters.R
\name{weightcalculatebis}
\alias{weightcalculatebis}
\title{Title}
\usage{
weightcalculatebis(X, mv, centers, clustering, gamma)
}
\arguments{
\item{X}{}

\item{mv}{}

\item{centers}{}

\item{clustering}{}

\item{gamma}{}
}
\description{
Title
}
