% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregHierarchiqueProbapost.R
\name{weightcalculateprobapost}
\alias{weightcalculateprobapost}
\title{Calculation of multi-view weights for hard clusters}
\usage{
weightcalculateprobapost(X, mv, centers, probapost, gamma)
}
\arguments{
\item{X}{Matrix of multi-view data, where the first view corresponds to the 
principal data used to obtain the partition in \code{cluster.init}}

\item{mv}{Vector corresponding to the size of each data view. The sum of \code{mv} should 
correspond to the number of columns in \code{X}.}

\item{centers}{Matrix of cluster centers}

\item{probapost}{Matrix of conditional probabilities of cluster membership for \code{X}}

\item{gamma}{??}
}
\description{
Calculation of multi-view weights for hard clusters
}
