% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPlot.R
\name{maskmeans_plot}
\alias{maskmeans_plot}
\title{Plot results of the multi-view aggregation/splitting K-means algorithm}
\usage{
maskmeans_plot(obj, type = c("dendrogram", "heights", "weights_line",
  "weights", "criterion", "tree", "tree_perClusterWeights"),
  tree_type = "final_K", mv_names = NULL, ...)
}
\arguments{
\item{obj}{Object of class \code{"maskmeans"} resulting from a call to the \code{maskemans}
function}

\item{type}{Graphic to be produced: one or more of \code{c("dendrogram", "heights", 
"weights_line", "weights", "criterion", "tree", "tree_perClusterWeights", probapost_boxplot")}.}

\item{tree_type}{Either \code{"final_K"}, \code{"all"}, or a numerical value for the final
number of clusters for plots of type \code{"tree"} or \code{"tree_perClusterWeights"}.}

\item{mv_names}{If desired, a vector of multiview names to be used for the 
\code{tree_perClusterWeights} plot}

\item{...}{Additional optional parameters.}
}
\value{
List of one or more ggplot2 objects.
}
\description{
Produce a variety of plots after running the multi-view aggregation or splitting
version of the K-means algorithm.
}
