% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPlot.R
\name{mv_plot}
\alias{mv_plot}
\title{Overview plot of multi-view data}
\usage{
mv_plot(
  mv_data,
  scale = TRUE,
  mv = NULL,
  mv_names = NULL,
  labels = NULL,
  include_default_caption = TRUE,
  ...
)
}
\arguments{
\item{mv_data}{Multi-view data, either in the form of a concatenated matrix (where the columns of the views have
been combined) or in the form of a list, where each element contains one of the views. In the former case, the argument
\code{mv} should also be provided to indicate the dimension of each of the views (\code{sum(mv)} should be equal to
the number of columns in the matrix). In the latter case, each matrix in the list should have observations (rows) sorted
in the same order, as matrices will be combined by columns within the function; note that \code{mv} will be detected 
automtically based on the number of columns in each matrix.}

\item{scale}{If \code{TRUE}, data will be scaled so that all variables are normalized and views are standardized according 
to their size.}

\item{mv}{(Optional unless \code{X} is a matrix.) If \code{X} is a matrix, vector 
corresponding to the size of each data view.}

\item{mv_names}{If desired, a vector of multiview names to be used for the plot.}

\item{labels}{If points should be colored by a hard cluster membership, a vector providing cluster assigments 
should be provided. If points should be colored by a soft cluster
membership (where the transparency of points reflects the maximum conditional probability of cluster membership), a 
matrix providing conditional probabilities of cluster membership should be provided.}

\item{include_default_caption}{If \code{TRUE}, include default caption.}

\item{...}{Additional optional parameters.}
}
\value{
A multi-facetted plot, with one facet per view. Univariate views are represented as densities, bivariate views
are represented as scatterplots, and multivariate views are represented by plotting scatterplots of the first two 
principal components.
}
\description{
Global graphical representation of multi-view data.
}
