% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatedData.R
\name{mv_simulate}
\alias{mv_simulate}
\title{Simulate multi-view data}
\usage{
mv_simulate(beta = 4, n = 100, K = 7, sigma = 1.5,
  beta_V6_multiplier = 1.5, ...)
}
\arguments{
\item{beta}{Multiplicative factor controlling the spread of values around the origin in the first (\code{beta}), 
second (\code{beta}), and sixth (\code{beta_V6_multiplier} x \code{beta}) views. Defaults to 4.}

\item{n}{Number of observations per cluster for simulations. Defaults to 100.}

\item{K}{Number of clusters in the first view for simulations. As this view is made up
of a single cluster at the origin surrounded by evenly spaced clusters in a circular pattern around
it, this number should be odd. Defaults to 7.}

\item{sigma}{Variance of noise to be added to views 2, 4, 5, and 6.
Defaults to 1.5.}

\item{beta_V6_multiplier}{Multiplicative parameter for the spread of values around the origin in view 6 
(\code{beta_V6_multiplier} x \code{beta}). Defaults to 1.5.}

\item{...}{Additional optional parameters.}
}
\value{
\item{data }{Multi-view simulated data}
\item{labels }{Matrix of dimension \code{n} x \code{v}, where \code{n} is the number
of observations and \code{v} the number of views, representing the true labels used
to generate the data}
}
\description{
Generate simulated multi-view data to illustrate the agglomerative and splitting versions of the
multi-view K-means algorithm.
}
\examples{
set.seed(12345)
sim_1 <- mv_simulate(type = "D1")
sim_2 <- mv_simulate(type = "D2")
sim_3 <- mv_simulate(type = "D3")
sim_4 <- mv_simulate(type = "D4")
sim_5 <- mv_simulate(type = "D5")
sim_6a <- mv_simulate(type = "D6")
sim_6b <- mv_simulate(type = "D6", delta=7, n=200, K=5, sigma=0.5)
}
