% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPlot.R
\name{PlotResHMV}
\alias{PlotResHMV}
\title{Title}
\usage{
PlotResHMV(Res, dendro = T, heights = T, weights = T, CRIT = T)
}
\arguments{
\item{Res}{}

\item{dendro}{}

\item{heights}{}

\item{weights}{}

\item{CRIT}{}
}
\description{
Title
}
