% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregHierarchique.R, R/SplittingClusters.R
\name{weightcalculate}
\alias{weightcalculate}
\alias{weightcalculate}
\title{Calculation of multi-view weights for hard clusters}
\usage{
weightcalculate(X, mv, centers, cluster, gamma)

weightcalculate(X, mv, centers, cluster, gamma)
}
\arguments{
\item{X}{}

\item{mv}{}

\item{centers}{Matrix of cluster centers}

\item{cluster}{Vector of cluster assignments}

\item{gamma}{}

\item{centers}{}

\item{cluster}{}
}
\description{
Calculation of multi-view weights for hard clusters

Title
}
